requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
crispy-bootstrap5==0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.5
django-crispy-forms==1.14.0
django-formtools==2.4
django-import-export==3.0.2
django-jazzmin==2.6.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.1.0
pathspec==0.10.3
platformdirs==2.6.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
sqlparse==0.4.3
stripe==5.0.0
tablib==3.3.0
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
